Fixed a bug, where if using profiler, after an hour of device ontime it would overflow the all buffers storing micros() data

+++ lib/Notifications & Profiler/profiler.h
@@ -8,7 +8,7 @@
 class Profiler
 {
 private:
-    uint32_t loop_start_time, function_start_time, end_function_time, end_loop_time;
+    uint64_t loop_start_time, function_start_time, end_function_time, end_loop_time;
     boolean enable_loop_measure;
 public:
     // @brief Calling this will begin measuring the delta time of a loop in milliseconds.
